refactored add_patient_form and fixed issues found on the way
